+++ packages/faceted-search/src/components/AddFacetPopover/AddFacetPopoverRow/index.js
@@ -1,0 +1,2 @@
+export * from './AddFacetPopoverRowButton';
+export * from './AddFacetPopoverRowItem';