+++ packages/faceted-search/src/components/AddFacetPopover/AddFacetPopver.component.test.js
@@ -1,7 +1,6 @@
-/* eslint-disable import/no-extraneous-dependencies */
 import React from 'react';
-import { mount } from 'enzyme';
 import { act } from 'react-dom/test-utils';
+import { mount } from 'enzyme';
 import { AddFacetPopover } from './AddFacetPopover.component';
 import getDefaultT from '../../translate';
 
@@ -65,6 +64,10 @@ describe('AddFacetPopover', () => {
 			},
 		},
 	];
+
+	const getRowButtons = wrapper =>
+		wrapper.find('div.tc-add-facet-popover-row-container').find('button');
+
 	it('should render the html output', () => {
 		// Given
 		const props = {
@@ -98,7 +101,9 @@ describe('AddFacetPopover', () => {
 		});
 		wrapper.update();
 		expect(wrapper.find('input').first().prop('value')).toBe('connection');
-		expect(wrapper.find('button[aria-label="Connection name"]')).toHaveLength(1);
+		const rowButton = getRowButtons(wrapper);
+		expect(rowButton).toHaveLength(1);
+		expect(rowButton.text()).toBe('Connection name');
 	});
 	it('should reset the badge rows when the filter is reset', () => {
 		// Given
@@ -112,14 +117,14 @@ describe('AddFacetPopover', () => {
 		};
 		// When
 		const wrapper = mount(<AddFacetPopover {...props} />);
+		expect(getRowButtons(wrapper)).toHaveLength(2);
 		// Then
 		act(() => {
 			wrapper.find('button[aria-label="Remove filter"]').first().simulate('mouseDown');
 		});
 		wrapper.update();
 		expect(wrapper.find('input').first().prop('value')).toBe('');
-		expect(wrapper.find('button[aria-label="Name"]')).toHaveLength(1);
-		expect(wrapper.find('button[aria-label="Connection name"]')).toHaveLength(1);
+		expect(getRowButtons(wrapper)).toHaveLength(3);
 	});
 	it('should return the badge definition when click on a row', () => {
 		// Given
@@ -133,9 +138,9 @@ describe('AddFacetPopover', () => {
 		};
 		// When
 		const wrapper = mount(<AddFacetPopover {...props} />);
-		wrapper.find('button[aria-label="Name"]').simulate('click');
+		getRowButtons(wrapper).first().simulate('click');
 		// Then
-		expect(onClick).toHaveBeenNthCalledWith(1, onClick.mock.calls[0][0], badgesDefinitions[0]);
+		expect(onClick).toHaveBeenNthCalledWith(1, onClick.mock.calls[0][0], badgesDefinitions[1]);
 	});
 	it('should render the category row', () => {
 		// Given
@@ -149,12 +154,10 @@ describe('AddFacetPopover', () => {
 		// When
 		const wrapper = mount(<AddFacetPopover {...props} />);
 		// Then
-		expect(wrapper.find('button[aria-label="Custom attributes"]')).toHaveLength(1);
-		expect(wrapper.find('.tc-add-facet-popover-screen')).toHaveLength(2);
-		expect(wrapper.find('.screen-category')).toHaveLength(1);
-		expect(wrapper.find('.screen-move')).toHaveLength(0);
+		expect(wrapper.find('.tc-add-facet-popover-row-button')).toHaveLength(3);
+		expect(wrapper.find('.tc-add-facet-popover-row-button-category')).toHaveLength(1);
 	});
-	it('should display the hidding category screen when click on a category row', () => {
+	it('should display the hidden category screen when click on a category row', () => {
 		// Given
 		const props = {
 			badges,
@@ -165,9 +168,9 @@ describe('AddFacetPopover', () => {
 		};
 		// When
 		const wrapper = mount(<AddFacetPopover {...props} />);
-		wrapper.find('button[aria-label="Custom attributes"]').simulate('click');
+		getRowButtons(wrapper).at(1).simulate('click'); // click on "Custom attributes"
 		// Then
-		expect(wrapper.find('.screen-move')).toHaveLength(2);
+		expect(wrapper.find('.tc-add-facet-popover-header-category')).toHaveLength(1);
 	});
 	it('should render an empty state when filter return no result', () => {
 		// Given
@@ -188,13 +191,13 @@ describe('AddFacetPopover', () => {
 				.simulate('change', { target: { value: 'aaaaaaaaaa' } });
 		});
 		wrapper.update();
-		expect(wrapper.find('button.tc-add-facet-popover-row')).toHaveLength(0);
+		expect(wrapper.find('button.tc-add-facet-popover-row-button')).toHaveLength(0);
 		expect(wrapper.find('span.tc-add-facet-popover-filter-empty').first()).toHaveLength(1);
 		expect(wrapper.find('span.tc-add-facet-popover-filter-empty').first().text()).toBe(
 			'No result found',
 		);
 	});
-	it('should render an disabled row if badgePerFacet is exceeded', () => {
+	it('should render a disabled row if badgePerFacet is exceeded', () => {
 		// Given
 		const props = {
 			badges: [
@@ -224,7 +227,8 @@ describe('AddFacetPopover', () => {
 		// When
 		const wrapper = mount(<AddFacetPopover {...props} />);
 		// Then
-		expect(wrapper.find('button[aria-label="Connection name"]').prop('disabled')).toBe(true);
+		expect(wrapper.find('button.tc-add-facet-popover-row-button')).toHaveLength(2);
+		expect(wrapper.find('div.tc-add-facet-popover-row-disabled')).toHaveLength(1);
 	});
 	it('should not render an empty label badge', () => {
 		// Given
@@ -271,11 +275,10 @@ describe('AddFacetPopover', () => {
 		const wrapper = mount(<AddFacetPopover {...props} />);
 
 		// Then
-		expect(wrapper.find('.tc-add-facet-popover-row-text').at(0).text()).toEqual('Connection name');
-		expect(wrapper.find('.tc-add-facet-popover-row-text').at(1).text()).toEqual(
-			'Custom attributes',
-		);
-		expect(wrapper.find('.tc-add-facet-popover-row-text').at(2).text()).toEqual('Name');
+		const rowButtons = getRowButtons(wrapper);
+		expect(rowButtons.at(0).text()).toEqual('Connection name');
+		expect(rowButtons.at(1).text()).toEqual('Custom attributes');
+		expect(rowButtons.at(2).text()).toEqual('Name');
 	});
 	it('should not sort if null is provided as comparator', () => {
 		// Given
@@ -291,10 +294,9 @@ describe('AddFacetPopover', () => {
 		const wrapper = mount(<AddFacetPopover {...props} />);
 
 		// Then
-		expect(wrapper.find('.tc-add-facet-popover-row-text').at(0).text()).toEqual('Name');
-		expect(wrapper.find('.tc-add-facet-popover-row-text').at(1).text()).toEqual('Connection name');
-		expect(wrapper.find('.tc-add-facet-popover-row-text').at(2).text()).toEqual(
-			'Custom attributes',
-		);
+		const rowButtons = getRowButtons(wrapper);
+		expect(rowButtons.at(0).text()).toEqual('Name');
+		expect(rowButtons.at(1).text()).toEqual('Connection name');
+		expect(rowButtons.at(2).text()).toEqual('Custom attributes');
 	});
 });