+++ packages/faceted-search/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 7.0.0
+
+### Major Changes
+
+- e5aa63ef6: Refactor with design system components and design tokens
+
+  Breaking: removed `FacetedSearchIcon` component
+
 ## 6.0.1
 
 ### Patch Changes